+++ mayan/apps/sources/models.py
@@ -8,19 +8,28 @@ import json
 import logging
 import os
 import poplib
+import subprocess
 
+import sh
 import yaml
+
+try:
+    scanimage = sh.Command('/usr/bin/scanimage')
+except sh.CommandNotFound:
+    scanimage = None
 
 from django.core.exceptions import ValidationError
 from django.core.files import File
 from django.core.files.base import ContentFile
 from django.db import models, transaction
 from django.utils.encoding import python_2_unicode_compatible
+from django.utils.timezone import now
 from django.utils.translation import ugettext_lazy as _
 
 from model_utils.managers import InheritanceManager
 
 from common.compressed_files import CompressedFile, NotACompressedFile
+from common.utils import TemporaryFile
 from converter.literals import DIMENSION_SEPARATOR
 from converter.models import Transformation
 from djcelery.models import PeriodicTask, IntervalSchedule
@@ -30,15 +39,20 @@ from metadata.api import save_metadata_list, set_bulk_metadata
 from metadata.models import MetadataType
 from tags.models import Tag
 
-from .classes import Attachment, SourceUploadedFile, StagingFile
+from .classes import Attachment, PseudoFile, SourceUploadedFile, StagingFile
+from .exceptions import SourceException
 from .literals import (
     DEFAULT_INTERVAL, DEFAULT_POP3_TIMEOUT, DEFAULT_IMAP_MAILBOX,
-    DEFAULT_METADATA_ATTACHMENT_NAME, SOURCE_CHOICES, SOURCE_CHOICE_STAGING,
-    SOURCE_CHOICE_WATCH, SOURCE_CHOICE_WEB_FORM,
-    SOURCE_INTERACTIVE_UNCOMPRESS_CHOICES, SOURCE_UNCOMPRESS_CHOICES,
-    SOURCE_UNCOMPRESS_CHOICE_N, SOURCE_UNCOMPRESS_CHOICE_Y,
-    SOURCE_CHOICE_EMAIL_IMAP, SOURCE_CHOICE_EMAIL_POP3
+    DEFAULT_METADATA_ATTACHMENT_NAME, SCANNER_ADF_MODE_CHOICES,
+    SCANNER_ADF_MODE_SIMPLEX, SCANNER_MODE_COLOR, SCANNER_MODE_CHOICES,
+    SCANNER_SOURCE_CHOICES, SCANNER_SOURCE_FLATBED,
+    SOURCE_CHOICES, SOURCE_CHOICE_STAGING, SOURCE_CHOICE_WATCH,
+    SOURCE_CHOICE_WEB_FORM, SOURCE_INTERACTIVE_UNCOMPRESS_CHOICES,
+    SOURCE_UNCOMPRESS_CHOICES, SOURCE_UNCOMPRESS_CHOICE_N,
+    SOURCE_UNCOMPRESS_CHOICE_Y, SOURCE_CHOICE_EMAIL_IMAP,
+    SOURCE_CHOICE_EMAIL_POP3, SOURCE_CHOICE_SANE_SCANNER,
 )
+from .settings import setting_scanimage_path
 
 logger = logging.getLogger(__name__)
 
@@ -153,6 +167,106 @@ class InteractiveSource(Source):
         verbose_name_plural = _('Interactive sources')
 
 
+class SaneScanner(InteractiveSource):
+    can_compress = False
+    is_interactive = True
+    source_type = SOURCE_CHOICE_SANE_SCANNER
+
+    device_name = models.CharField(
+        max_length=255,
+        help_text=_('Device name as returned by the SANE backend.'),
+        verbose_name=_('Device name')
+    )
+    mode = models.CharField(
+        blank=True, choices=SCANNER_MODE_CHOICES, default=SCANNER_MODE_COLOR,
+        help_text=_(
+            'Selects the scan mode (e.g., lineart, monochrome, or color). '
+            'If this option is not supported by your scanner, leave it blank.'
+        ), max_length=16, verbose_name=_('Mode')
+    )
+    resolution = models.PositiveIntegerField(
+        blank=True, help_text=_(
+            'Sets the resolution of the scanned image in DPI (dots per inch). '
+            'Typical value is 200. If this option is not supported by your '
+            'scanner, leave it blank.'
+        ), verbose_name=_('Resolution')
+    )
+    source = models.CharField(
+        blank=True, choices=SCANNER_SOURCE_CHOICES,
+        default=SCANNER_SOURCE_FLATBED, help_text=_(
+            'Selects the scan source (such as a document-feeder). If this '
+            'option is not supported by your scanner, leave it blank.'
+        ), max_length=32,
+        verbose_name=_('Paper source')
+    )
+    adf_mode = models.CharField(
+        blank=True, choices=SCANNER_ADF_MODE_CHOICES,
+        default=SCANNER_ADF_MODE_SIMPLEX, help_text=_(
+            'Selects the document feeder mode (simplex/duplex). If this '
+            'option is not supported by your scanner, leave it blank.'
+        ), max_length=16, verbose_name=_('ADF mode')
+    )
+
+    class Meta:
+        verbose_name = _('SANE Scanner')
+        verbose_name_plural = _('SANE Scanners')
+
+    def clean_up_upload_file(self, upload_file_object):
+        pass
+
+    def get_upload_file_object(self, form_data):
+        temporary_file_object = TemporaryFile()
+        command_line = [
+            setting_scanimage_path.value, '-d', self.device_name,
+            '--format', 'tiff',
+        ]
+
+        if self.resolution:
+            command_line.extend(
+                ['--resolution', '{}'.format(self.resolution)]
+            )
+
+        if self.mode:
+            command_line.extend(
+                ['--mode', self.mode]
+            )
+
+        if self.source:
+            command_line.extend(
+                ['--source', self.source]
+            )
+
+        if self.adf_mode:
+            command_line.extend(
+                ['--adf-mode', self.adf_mode]
+            )
+
+        filestderr = TemporaryFile()
+
+        try:
+            logger.debug('Scan command line: %s', command_line)
+            subprocess.check_call(
+                command_line, stdout=temporary_file_object, stderr=filestderr
+            )
+        except subprocess.CalledProcessError:
+            filestderr.seek(0)
+            error_message = filestderr.read()
+            logger.error(
+                'Exception while scanning from source:%s ; %s', self,
+                error_message
+            )
+
+            message = _('Error while scanning; %s') % error_message
+            self.logs.create(message=message)
+            raise SourceException(message)
+        else:
+            return SourceUploadedFile(
+                source=self, file=PseudoFile(
+                    file=temporary_file_object, name='scan {}'.format(now())
+                )
+            )
+
+
 class StagingFolderSource(InteractiveSource):
     """
     The Staging folder source is interactive but instead of displaying an
@@ -169,6 +283,7 @@ class StagingFolderSource(InteractiveSource):
     convert the scanned files into Mayan EDMS documents. Staging folders are
     useful when many users share a few networked scanners.
     """
+    can_compress = True
     is_interactive = True
     source_type = SOURCE_CHOICE_STAGING
 
@@ -257,6 +372,7 @@ class WebFormSource(InteractiveSource):
     documents from their own computers such as when each user has their own
     scanner.
     """
+    can_compress = True
     is_interactive = True
     source_type = SOURCE_CHOICE_WEB_FORM