+++ mayan/apps/sources/views.py
@@ -24,17 +24,13 @@ from documents.tasks import task_upload_new_version
 from metadata.api import decode_metadata_from_url
 from navigation import Link
 
+from .exceptions import SourceException
 from .forms import (
-    NewDocumentForm, NewVersionForm, WebFormUploadForm,
-    WebFormUploadFormHTML5
+    NewDocumentForm, NewVersionForm, WebFormUploadForm, WebFormUploadFormHTML5
 )
-from .literals import (
-    SOURCE_CHOICE_STAGING, SOURCE_CHOICE_WEB_FORM,
-    SOURCE_UNCOMPRESS_CHOICE_ASK,
-    SOURCE_UNCOMPRESS_CHOICE_Y
-)
+from .literals import SOURCE_UNCOMPRESS_CHOICE_ASK, SOURCE_UNCOMPRESS_CHOICE_Y
 from .models import (
-    InteractiveSource, Source, StagingFolderSource, WebFormSource
+    InteractiveSource, Source, SaneScanner, StagingFolderSource
 )
 from .permissions import (
     permission_sources_setup_create, permission_sources_setup_delete,
@@ -88,27 +84,10 @@ class UploadBaseView(MultiFormView):
 
     @staticmethod
     def get_active_tab_links(document=None):
-        tab_links = []
-
-        web_forms = WebFormSource.objects.filter(enabled=True)
-        for web_form in web_forms:
-            tab_links.append(
-                UploadBaseView.get_tab_link_for_source(web_form, document)
-            )
-
-        staging_folders = StagingFolderSource.objects.filter(enabled=True)
-        for staging_folder in staging_folders:
-            tab_links.append(
-                UploadBaseView.get_tab_link_for_source(
-                    staging_folder, document
-                )
-            )
-
-        return {
-            'tab_links': tab_links,
-            SOURCE_CHOICE_WEB_FORM: web_forms,
-            SOURCE_CHOICE_STAGING: staging_folders,
-        }
+        return [
+            UploadBaseView.get_tab_link_for_source(source, document)
+            for source in InteractiveSource.objects.filter(enabled=True).select_subclasses()
+        ]
 
     def dispatch(self, request, *args, **kwargs):
         if 'source_id' in kwargs:
@@ -163,6 +142,16 @@ class UploadBaseView(MultiFormView):
                         }
                     },
                 ]
+        elif isinstance(self.source, SaneScanner):
+            subtemplates_list.append({
+                'name': 'sources/upload_multiform_subtemplate.html',
+                'context': {
+                    'forms': context['forms'],
+                    'is_multipart': True,
+                    'title': _('Document properties'),
+                    'submit_label': _('Scan'),
+                },
+            })
         else:
             subtemplates_list.append({
                 'name': 'sources/upload_multiform_subtemplate.html',
@@ -173,8 +162,8 @@ class UploadBaseView(MultiFormView):
                 },
             })
 
-        menu_facet.bound_links['sources:upload_interactive'] = self.tab_links['tab_links']
-        menu_facet.bound_links['sources:upload_version'] = self.tab_links['tab_links']
+        menu_facet.bound_links['sources:upload_interactive'] = self.tab_links
+        menu_facet.bound_links['sources:upload_version'] = self.tab_links
 
         context.update({
             'subtemplates_list': subtemplates_list,
@@ -206,57 +195,63 @@ class UploadInteractiveView(UploadBaseView):
         ).dispatch(request, *args, **kwargs)
 
     def forms_valid(self, forms):
-        if self.source.uncompress == SOURCE_UNCOMPRESS_CHOICE_ASK:
-            expand = forms['source_form'].cleaned_data.get('expand')
-        else:
-            if self.source.uncompress == SOURCE_UNCOMPRESS_CHOICE_Y:
-                expand = True
+        if self.source.can_compress:
+            if self.source.uncompress == SOURCE_UNCOMPRESS_CHOICE_ASK:
+                expand = forms['source_form'].cleaned_data.get('expand')
             else:
-                expand = False
-
-        uploaded_file = self.source.get_upload_file_object(
-            forms['source_form'].cleaned_data
-        )
-
-        shared_uploaded_file = SharedUploadedFile.objects.create(
-            file=uploaded_file.file
-        )
-
-        label = None
-
-        if 'document_type_available_filenames' in forms['document_form'].cleaned_data:
-            if forms['document_form'].cleaned_data['document_type_available_filenames']:
-                label = forms['document_form'].cleaned_data['document_type_available_filenames'].filename
-
-        if not self.request.user.is_anonymous():
-            user_id = self.request.user.pk
+                if self.source.uncompress == SOURCE_UNCOMPRESS_CHOICE_Y:
+                    expand = True
+                else:
+                    expand = False
         else:
-            user_id = None
+            expand = False
 
         try:
-            self.source.clean_up_upload_file(uploaded_file)
-        except Exception as exception:
-            messages.error(self.request, exception)
-
-        task_source_handle_upload.apply_async(kwargs=dict(
-            description=forms['document_form'].cleaned_data.get('description'),
-            document_type_id=self.document_type.pk,
-            expand=expand,
-            label=label,
-            language=forms['document_form'].cleaned_data.get('language'),
-            metadata_dict_list=decode_metadata_from_url(self.request.GET),
-            shared_uploaded_file_id=shared_uploaded_file.pk,
-            source_id=self.source.pk,
-            tag_ids=self.request.GET.getlist('tags'),
-            user_id=user_id,
-        ))
-        messages.success(
-            self.request,
-            _(
-                'New document queued for uploaded and will be available '
-                'shortly.'
+            uploaded_file = self.source.get_upload_file_object(
+                forms['source_form'].cleaned_data
             )
-        )
+        except SourceException as exception:
+            messages.error(self.request, exception)
+        else:
+            shared_uploaded_file = SharedUploadedFile.objects.create(
+                file=uploaded_file.file
+            )
+
+            label = None
+
+            if 'document_type_available_filenames' in forms['document_form'].cleaned_data:
+                if forms['document_form'].cleaned_data['document_type_available_filenames']:
+                    label = forms['document_form'].cleaned_data['document_type_available_filenames'].filename
+
+            if not self.request.user.is_anonymous():
+                user_id = self.request.user.pk
+            else:
+                user_id = None
+
+            try:
+                self.source.clean_up_upload_file(uploaded_file)
+            except Exception as exception:
+                messages.error(self.request, exception)
+
+            task_source_handle_upload.apply_async(kwargs=dict(
+                description=forms['document_form'].cleaned_data.get('description'),
+                document_type_id=self.document_type.pk,
+                expand=expand,
+                label=label,
+                language=forms['document_form'].cleaned_data.get('language'),
+                metadata_dict_list=decode_metadata_from_url(self.request.GET),
+                shared_uploaded_file_id=shared_uploaded_file.pk,
+                source_id=self.source.pk,
+                tag_ids=self.request.GET.getlist('tags'),
+                user_id=user_id,
+            ))
+            messages.success(
+                self.request,
+                _(
+                    'New document queued for uploaded and will be available '
+                    'shortly.'
+                )
+            )
 
         return HttpResponseRedirect(
             '{}?{}'.format(
@@ -266,12 +261,15 @@ class UploadInteractiveView(UploadBaseView):
         )
 
     def create_source_form_form(self, **kwargs):
+        if hasattr(self.source, 'uncompress'):
+            show_expand = self.source.uncompress == SOURCE_UNCOMPRESS_CHOICE_ASK
+        else:
+            show_expand = False
+
         return self.get_form_classes()['source_form'](
             prefix=kwargs['prefix'],
             source=self.source,
-            show_expand=(
-                self.source.uncompress == SOURCE_UNCOMPRESS_CHOICE_ASK
-            ),
+            show_expand=show_expand,
             data=kwargs.get('data', None),
             files=kwargs.get('files', None),
         )
@@ -301,7 +299,7 @@ class UploadInteractiveView(UploadBaseView):
         context['title'] = _(
             'Upload a local document from source: %s'
         ) % self.source.label
-        if not isinstance(self.source, StagingFolderSource):
+        if not isinstance(self.source, StagingFolderSource) and not isinstance(self.source, SaneScanner):
             context['subtemplates_list'][0]['context'].update(
                 {
                     'form_action': '{}?{}'.format(
@@ -349,38 +347,42 @@ class UploadInteractiveVersionView(UploadBaseView):
         ).dispatch(request, *args, **kwargs)
 
     def forms_valid(self, forms):
-        uploaded_file = self.source.get_upload_file_object(
-            forms['source_form'].cleaned_data
-        )
-
-        shared_uploaded_file = SharedUploadedFile.objects.create(
-            file=uploaded_file.file
-        )
-
         try:
-            self.source.clean_up_upload_file(uploaded_file)
-        except Exception as exception:
-            messages.error(self.request, exception)
-
-        if not self.request.user.is_anonymous():
-            user_id = self.request.user.pk
-        else:
-            user_id = None
-
-        task_upload_new_version.apply_async(kwargs=dict(
-            shared_uploaded_file_id=shared_uploaded_file.pk,
-            document_id=self.document.pk,
-            user_id=user_id,
-            comment=forms['document_form'].cleaned_data.get('comment')
-        ))
-
-        messages.success(
-            self.request,
-            _(
-                'New document version queued for uploaded and will be '
-                'available shortly.'
+            uploaded_file = self.source.get_upload_file_object(
+                forms['source_form'].cleaned_data
             )
-        )
+        except SourceException as exception:
+            messages.error(self.request, exception)
+        else:
+            shared_uploaded_file = SharedUploadedFile.objects.create(
+                file=uploaded_file.file
+            )
+
+            try:
+                self.source.clean_up_upload_file(uploaded_file)
+            except Exception as exception:
+                messages.error(self.request, exception)
+
+            if not self.request.user.is_anonymous():
+                user_id = self.request.user.pk
+            else:
+                user_id = None
+
+            task_upload_new_version.apply_async(kwargs=dict(
+                shared_uploaded_file_id=shared_uploaded_file.pk,
+                document_id=self.document.pk,
+                user_id=user_id,
+                comment=forms['document_form'].cleaned_data.get('comment')
+            ))
+
+            messages.success(
+                self.request,
+                _(
+                    'New document version queued for uploaded and will be '
+                    'available shortly.'
+                )
+            )
+
         return HttpResponseRedirect(
             reverse(
                 'documents:document_version_list', args=(self.document.pk,)