+++ mayan/apps/permissions/urls.py
@@ -30,5 +30,4 @@ api_urls = [
     url(r'^permissions/$', APIPermissionList.as_view(), name='permission-list'),
     url(r'^roles/$', APIRoleListView.as_view(), name='role-list'),
     url(r'^roles/(?P<pk>[0-9]+)/$', APIRoleView.as_view(), name='role-detail'),
-    url(r'^$', APIPermissionList.as_view(), name='permission-list'),
 ]