+++ HISTORY.rst
@@ -3,6 +3,7 @@ XX (2017-XX-XX)
 - Add Django-mathfilters.
 - Improve render of documents with no pages.
 - Add SANE scanner document source.
+- Added PDF orientation detection.
 
 2.3 (2017-06-08)
 ================