+++ HISTORY.rst
@@ -3,7 +3,8 @@ XX (2017-XX-XX)
 - Add Django-mathfilters.
 - Improve render of documents with no pages.
 - Add SANE scanner document source.
-- Added PDF orientation detection.
+- Added PDF orientation detection. GitLab issue #387.
+- Fix repeated permission list API URL. GitLab issue #389.
 
 2.3 (2017-06-08)
 ================