+++ HISTORY.rst
@@ -5,6 +5,7 @@ XX (2017-XX-XX)
 - Add SANE scanner document source.
 - Added PDF orientation detection. GitLab issue #387.
 - Fix repeated permission list API URL. GitLab issue #389.
+- Fix role creation API endpoint not returning id. GitLab issue #390.
 
 2.3 (2017-06-08)
 ================