+++ HISTORY.rst
@@ -7,6 +7,7 @@ XX (2017-XX-XX)
 - Fix repeated permission list API URL. GitLab issue #389.
 - Fix role creation API endpoint not returning id. GitLab issue #390.
 - Make tags, metadata types and cabinets searchable via the dynamic search API. GitLab issue #344.
+- Add support for updating configuration options from environment variables.
 
 2.3 (2017-06-08)
 ================