+++ HISTORY.rst
@@ -8,7 +8,8 @@ XX (2017-XX-XX)
 - Fix role creation API endpoint not returning id. GitLab issue #390.
 - Make tags, metadata types and cabinets searchable via the dynamic search API. GitLab issue #344.
 - Add support for updating configuration options from environment variables.
-
+- Add purgelocks management command. GitLab issue #221.
+ 
 2.3 (2017-06-08)
 ================
 - Allow for bigger indexing expression templates.